Tempo.  It should now be trivial to implement changes of tempo from the GUI i.e. the Sequencer works with the tempo rather than ignoring it.  All of this should go to help timing for the MIDI export code as well.  A default tempo is assigned to a loaded file if no tempo is found in it.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@935 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ base/Composition.h
@@ -82,10 +82,13 @@ public:
     // MIDI has a tempo meta-event which can change arbitrarily often,
     // and notation has metronome events that can also appear more than
     // once (and indeed per Track, rather than per Composition).
-    // (We perhaps do need a per-Composition base tempo, though.)
-    unsigned int getTempo() const { return m_tempo; }
-    void setTempo(const int &tempo) { m_tempo = tempo; }
-
+    // (We perhaps do need a per-Composition base tempo, though.) [cc]
+    //
+    // Tempo here is only our current Transport tempo which we use on
+    // the GUI and is sent to the Sequencer.
+    //
+    double getTempo() const { return m_tempo; }
+    void setTempo(const double &tempo) { m_tempo = tempo; }
 
     const timeT& getPosition() { return m_position; }
     void setPosition(const timeT& position) { m_position = position; }
@@ -101,7 +104,7 @@ protected:
     Track m_timeReference; // contains time signature events &c
 
     unsigned int m_nbTicksPerBar; //!!! must lose this
-    unsigned int m_tempo;
+    double m_tempo;
 
     timeT m_position;
